Separated out the logic for file reading in hook
-> Added custom hook for source and target files reading
-> Changed the heading and fields to show in FieldSelector component
on the basis of props recieved
