+++ src/components/FieldSelector/FieldSelector.jsx
@@ -3,10 +3,18 @@ import FieldBox from "../FieldBox/FieldBox";
 import styles from "./FieldSelector.module.css";
 import { useSelector } from "react-redux";
 
-const FieldSelector = () => {
-  const sourceField = useSelector((state) => state.fields.sourceFields);
+const FieldSelector = (props) => {
+  let typeOfField;
 
-  const content = sourceField.map((data, idx) => (
+  const { type, heading } = props;
+
+  if (type === "source") {
+    typeOfField = useSelector((state) => state.fields.sourceFields);
+  } else {
+    typeOfField = useSelector((state) => state.fields.targetFields);
+  }
+
+  const content = typeOfField.map((data, idx) => (
     <FieldBox
       fileName={data.name}
       fields={data.result}
@@ -17,7 +25,7 @@ const FieldSelector = () => {
 
   return (
     <article>
-      <h3>Source Field Selection</h3>
+      <h3>{heading}</h3>
       <div className={styles.container}>{content}</div>
     </article>
   );
